more information on sequences used for phylogenies
This commit output a CSV and FASTA file with not only the sequence used for the phylogenies, but also all the sequences used for without filtering to those observed multiple times. Previously, just the sequences filtered for multiple observations were output (in `results/phylogenetics/all_alignment.csv`). Now the ones observed just once are also output (in `results/phylogenetics/all_alignment_no_filter_rare.csv`). Note that the CSV files with the sequence information is tracked in the GitHub repo, but not the raw sequences themselves due to GitHub data sharing. However, the sequences themselves can be re-constructed by adding the mutations back to the reference genome for *proCoV2* tracked at `results/ref_genome/ref_genome.fa`.

This commit is to provide data for analyses that @matsen and @WSDeWitt want to do.