add analysis of SRR11313490 and SRR11313499
After posting of the initial pre-print, I was provided information on how to obtain the FASTQs for these runs unavailable on the Google Cloud from the following three researchers:
  - Andrew Fire (Stanford, https://en.wikipedia.org/wiki/Andrew_Fire)
  - Theo Sanderson (Crick Institute, https://theo.io/)
  - Francisco de Asis (DRASTIC Research, https://drasticresearch.org/tag/francisco-de-asis/)

Adding the new data did not appreciably change any results because they were all from low-coverage samples for which it was still not possible to call much about the viral sequence. But it adds "completeness" to the analysis.